requirements.txt
google-adk>=1.26.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
firebase-admin>=6.5.0
aiosmtplib>=3.0.0
email-validator>=2.1.0
